# Get rid of MVCCValue.