… storage/engine/rocksdb/cockroach/storage/engine/mvcc.pb.cc
… storage/replica_test.go