Deletion tombstones are now indicated by an empty value. Not having to
wrap a Value with an MVCCValue gives a nice perf boost.
```
name                          old time/op    new time/op    delta
MVCCScan1Version1Row-8          5.99µs ±14%    6.11µs ±15%     ~           (p=0.796 n=10+10)
MVCCScan1Version10Rows-8        15.7µs ± 3%    15.6µs ± 1%     ~            (p=0.182 n=10+9)
MVCCScan1Version100Rows-8        176µs ±17%     144µs ± 1%  -18.13%         (p=0.000 n=9+10)
MVCCScan1Version1000Rows-8      1.49ms ± 3%    1.44ms ± 4%   -3.16%         (p=0.009 n=8+10)
MVCCScan10Versions1Row-8        8.65µs ±14%    5.91µs ± 8%  -31.72%        (p=0.000 n=10+10)
MVCCScan10Versions10Rows-8      22.0µs ± 2%    19.4µs ± 6%  -11.80%         (p=0.000 n=9+10)
MVCCScan10Versions100Rows-8      200µs ±13%     185µs ± 4%   -7.37%        (p=0.000 n=10+10)
MVCCScan10Versions1000Rows-8    1.89ms ± 5%    1.75ms ± 2%   -7.30%        (p=0.000 n=10+10)
MVCCGet1Version-8               6.16µs ±12%    6.00µs ±17%     ~           (p=0.436 n=10+10)
MVCCGet10Versions-8             7.54µs ± 5%    5.63µs ± 4%  -25.33%         (p=0.000 n=8+10)

name                          old speed      new speed      delta
MVCCScan1Version1Row-8         171MB/s ±12%   169MB/s ±13%     ~           (p=0.796 n=10+10)
MVCCScan1Version10Rows-8       651MB/s ± 3%   657MB/s ± 1%     ~            (p=0.182 n=10+9)
MVCCScan1Version100Rows-8      586MB/s ±19%   709MB/s ± 1%  +21.09%         (p=0.000 n=9+10)
MVCCScan1Version1000Rows-8     687MB/s ± 3%   710MB/s ± 4%   +3.28%         (p=0.009 n=8+10)
MVCCScan10Versions1Row-8       119MB/s ±13%   174MB/s ± 8%  +45.84%        (p=0.000 n=10+10)
MVCCScan10Versions10Rows-8     466MB/s ± 2%   529MB/s ± 6%  +13.48%         (p=0.000 n=9+10)
MVCCScan10Versions100Rows-8    514MB/s ±11%   553MB/s ± 4%   +7.70%        (p=0.000 n=10+10)
MVCCScan10Versions1000Rows-8   542MB/s ± 5%   585MB/s ± 3%   +7.83%        (p=0.000 n=10+10)
MVCCGet1Version-8              167MB/s ±12%   172MB/s ±15%     ~           (p=0.436 n=10+10)
MVCCGet10Versions-8            136MB/s ± 5%   182MB/s ± 4%  +33.97%         (p=0.000 n=8+10)
```

<!-- Reviewable:start -->
[<img src="https://reviewable.io/review_button.png" height=40 alt="Review on Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/3217)
<!-- Reviewable:end -->