@tschottdorf 2015-11-24
LGTM. Why do you think the speedup is so noticeable?