@petermattis 2015-11-24
@tschottdorf Not sure why the speedup is so large. It does remove an allocation from the inner scan loop. And it removes some of the marshaling. Still, the speedup is larger than expected.